+++ BlazorCore/Models/AppSettingsEntity.cs
@@ -1,5 +1,6 @@
 ﻿using BlazorCore.DAL;
 using BlazorCore.Utils;
+using FluentNHibernate.Automapping.Steps;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Authorization;
 using System;
@@ -23,32 +24,39 @@ namespace BlazorCore.Models
 
         public AppSettingsEntity() { }
 
-        public void Setup(AuthenticationStateProvider stateProvider, JsonAppSettingsEntity jsonAppSettings, HotKeys hotKeys)
+        public void Setup(AuthenticationStateProvider stateProvider, JsonAppSettingsEntity jsonAppSettings, HotKeys hotKeys,
+            [CallerMemberName] string memberName = "")
         {
-            if (jsonAppSettings != null)
+            lock (Locker)
             {
-                JsonAppSettings = jsonAppSettings;
-                CoreSettingsEntity appSettings = new(JsonAppSettings.Server, JsonAppSettings.Db, JsonAppSettings.Trusted,
-                    JsonAppSettings.Username, JsonAppSettings.Password);
-                DataAccess = new DataAccessEntity(appSettings);
-            }
-            if (hotKeys != null)
-                HotKeysItem = hotKeys;
-            if (stateProvider != null)
-                IdentityOpen(stateProvider);
+                if (jsonAppSettings != null && JsonAppSettings == null && CoreSettingsItem == null && DataAccess == null)
+                {
+                    JsonAppSettings = jsonAppSettings;
+                    CoreSettingsItem = new(JsonAppSettings.Server, JsonAppSettings.Db, JsonAppSettings.Trusted, JsonAppSettings.Username, JsonAppSettings.Password);
+                    DataAccess = new DataAccessEntity(CoreSettingsItem);
+                }
+                if (hotKeys != null)
+                {
+                    HotKeysItem?.DisposeAsync().ConfigureAwait(true);
+                    HotKeysItem = hotKeys;
+                }
+                
+                if (stateProvider != null)
+                    IdentityOpen(stateProvider);
 
-            // Debug log.
-            if (IsDebug)
-            {
-                Console.WriteLine("--------------------------------------------------------------------------------");
-                Console.WriteLine($"---------- {nameof(AppSettingsEntity)}.{nameof(Setup)} (for Debug mode) ---------- ");
-                Console.WriteLine($"{nameof(Memory)}: {Memory}");
-                Console.WriteLine($"{nameof(HotKeysItem)}: {HotKeysItem}");
-                Console.WriteLine($"{nameof(JsonAppSettings)}: {JsonAppSettings}");
-                Console.WriteLine($"{nameof(DataAccess.CoreSettings)}: {DataAccess.CoreSettings}");
-                Console.WriteLine($"{nameof(DataAccess)}: {DataAccess}");
-                Console.WriteLine($"{nameof(DataSource)}: {DataSource}");
-                Console.WriteLine("--------------------------------------------------------------------------------");
+                // Debug log.
+                if (IsDebug)
+                {
+                    Console.WriteLine("--------------------------------------------------------------------------------");
+                    Console.WriteLine($"---------- {nameof(AppSettingsEntity)}.{memberName} (for Debug mode) ---------- ");
+                    Console.WriteLine($"{nameof(Memory)}: {Memory}");
+                    Console.WriteLine($"{nameof(HotKeysItem)}: {HotKeysItem}");
+                    Console.WriteLine($"{nameof(JsonAppSettings)}: {JsonAppSettings}");
+                    Console.WriteLine($"{nameof(DataAccess.CoreSettings)}: {DataAccess.CoreSettings}");
+                    Console.WriteLine($"{nameof(DataAccess)}: {DataAccess}");
+                    Console.WriteLine($"{nameof(DataSource)}: {DataSource}");
+                    Console.WriteLine("--------------------------------------------------------------------------------");
+                }
             }
         }
 
@@ -56,8 +64,9 @@ namespace BlazorCore.Models
 
         #region Public and private fields and properties
 
-        public bool? IdentityAccessLevel { get; private set; }
-
+        public object Locker { get; set; } = new object();
+        public CoreSettingsEntity CoreSettingsItem { get; set; }
+        public IdentityEntity IdentityItem { get; private set; }
         public DataAccessEntity DataAccess { get; private set; }
         public DataSourceEntity DataSource { get; private init; } = new();
         public JsonAppSettingsEntity JsonAppSettings { get; private set; }
@@ -69,10 +78,10 @@ namespace BlazorCore.Models
         public int FontSizeHeader { get; set; }
         public int FontSize { get; set; }
         public int Delay => 5_000;
-        public string MemoryInfo => Memory != null
+        public string MemoryInfo => Memory != null && Memory.MemorySize != null &&
+            Memory.MemorySize.Physical != null 
             ? $"{LocalizationStrings.DeviceControl.MemoryUsed}: {Memory.MemorySize.Physical.MegaBytes:N0} MB  |  {UtilsDt.FormatCurDtRus(true)}"
             : $"{LocalizationStrings.DeviceControl.MemoryUsed}: - MB";
-        public string IdentityName { get; private set; }
         public bool IsChartSmooth { get; set; }
 
         public string SqlServerDescription => JsonAppSettings is { Server: { } }
@@ -100,37 +109,41 @@ namespace BlazorCore.Models
 
         #region Public and private methods - Authentication & identity
 
-        public void IdentityOpen(AuthenticationStateProvider stateProvider)
+        private void IdentityOpen(AuthenticationStateProvider stateProvider)
         {
             if (stateProvider != null)
             {
                 AuthenticationState authenticationState = stateProvider.GetAuthenticationStateAsync().Result;
                 IIdentity identity = authenticationState?.User?.Identity;
-                IdentityName = identity != null ? identity.Name : LocalizationStrings.Share.IdentityError;
-                SetUserAccessLevel(IdentityName);
+                if (IdentityItem == null)
+                    IdentityItem = new IdentityEntity(identity != null ? identity.Name : LocalizationStrings.Share.IdentityError);
+                else
+                    IdentityItem.Name = identity != null ? identity.Name : LocalizationStrings.Share.IdentityError;
+                SetUserAccessLevel();
             }
         }
 
-        private void SetUserAccessLevel(string userName,
-            [CallerFilePath] string filePath = "", [CallerLineNumber] int lineNumber = 0, [CallerMemberName] string memberName = "")
+        private void SetUserAccessLevel([CallerFilePath] string filePath = "", [CallerLineNumber] int lineNumber = 0, [CallerMemberName] string memberName = "")
         {
-            IdentityAccessLevel = null;
-            if (DataAccess == null)
-                return;
-            object[] objects = DataAccess.GetEntitiesNativeObject(SqlQueries.GetAccessUser(userName), filePath, lineNumber, memberName);
-            if (objects.Length == 1)
+            IdentityItem.IsAccess = false;
+            if (DataAccess != null)
             {
-                if (objects[0] is object[] { Length: 5 } item)
+                object[] objects = DataAccess.GetEntitiesNativeObject(SqlQueries.GetAccessUser(IdentityItem.Name), filePath, lineNumber, memberName);
+                if (objects.Length == 1)
                 {
-                    if (Guid.TryParse(Convert.ToString(item[0]), out _))
+                    if (objects[0] is object[] { Length: 5 } item)
                     {
-                        if (item[4] != null)
+                        if (Guid.TryParse(Convert.ToString(item[0]), out _))
                         {
-                            IdentityAccessLevel = Convert.ToBoolean(item[4]);
+                            if (item[4] != null)
+                            {
+                                IdentityItem.AccessLevel = Convert.ToBoolean(item[4]);
+                            }
                         }
                     }
                 }
             }
+            IdentityItem.IsAccess = true;
         }
 
         #endregion