+++ BlazorCore/Models/CoreSettingsEntity.cs
@@ -57,7 +57,7 @@ namespace BlazorCore.Models
 
         public override string ToString()
         {
-            var strTrusted = Trusted ? $"{nameof(Trusted)}: true." : $"{nameof(Username)}: {Username}. {nameof(Password)}: {Password}.";
+            string strTrusted = Trusted ? $"{nameof(Trusted)}: true." : $"{nameof(Username)}: {Username}. {nameof(Password)}: {Password}.";
             return $"{nameof(Server)}: { Server}. " +
                    $"{nameof(Db)}: {Db}. " +
                    $"{strTrusted}";