+++ BlazorDeviceControl/Shared/Section/Printers.razor.cs
@@ -93,7 +93,7 @@ namespace BlazorDeviceControl.Shared.Section
 
         private async Task ActionEditAsync(EnumTable table, BaseIdEntity item, BaseIdEntity parentEntity)
         {
-            if (AppSettings.IdentityAccessLevel != true)
+            if (AppSettings.IdentityItem.AccessLevel != true)
                 return;
 
             await RunTasksAsync(LocalizationStrings.DeviceControl.GetItemTitle(table), "", LocalizationStrings.Share.DialogResultFail, "",
@@ -107,7 +107,7 @@ namespace BlazorDeviceControl.Shared.Section
 
         private async Task ActionAddAsync(EnumTable table, BaseIdEntity entity, BaseIdEntity parentEntity)
         {
-            if (AppSettings.IdentityAccessLevel != true)
+            if (AppSettings.IdentityItem.AccessLevel != true)
                 return;
 
             await ActionAsync<BaseRazorEntity>(table, EnumTableAction.Add, entity, parentEntity).ConfigureAwait(true);
@@ -116,7 +116,7 @@ namespace BlazorDeviceControl.Shared.Section
 
         private async Task ActionCopyAsync(EnumTable table, BaseIdEntity entity, BaseIdEntity parentEntity)
         {
-            if (AppSettings.IdentityAccessLevel != true)
+            if (AppSettings.IdentityItem.AccessLevel != true)
                 return;
 
             await ActionAsync<BaseRazorEntity>(table, EnumTableAction.Copy, entity, parentEntity).ConfigureAwait(true);
@@ -125,7 +125,7 @@ namespace BlazorDeviceControl.Shared.Section
 
         private async Task ActionDeleteAsync(EnumTable table, BaseIdEntity entity, BaseIdEntity parentEntity)
         {
-            if (AppSettings.IdentityAccessLevel != true)
+            if (AppSettings.IdentityItem.AccessLevel != true)
                 return;
 
             await ActionAsync<BaseRazorEntity>(table, EnumTableAction.Delete, entity, parentEntity).ConfigureAwait(true);
@@ -134,7 +134,7 @@ namespace BlazorDeviceControl.Shared.Section
 
         private async Task ActionMarkedAsync(EnumTable table, BaseIdEntity entity, BaseIdEntity parentEntity)
         {
-            if (AppSettings.IdentityAccessLevel != true)
+            if (AppSettings.IdentityItem.AccessLevel != true)
                 return;
 
             await ActionAsync<BaseRazorEntity>(table, EnumTableAction.Marked, entity, parentEntity).ConfigureAwait(true);